refactor(cli): move template import functionality into Runner
- Removed standalone import.rs module
- Moved add_templates_in_renderer and build_templates_import_globset methods into Runner struct
- Updated imports in runner.rs to include required dependencies
- Maintained same functionality while improving code organization
